Getting saving of inline comments working, and starting to work on creating sections properly.
There's a dependency cycle in the UI for creating courses, sections, and registrations: you cannot create a registration for a professor without assigning them a section, but you cannot create a section without it having a professor.  The interim approach is to allow the course-creation and course-modification pages to allow specifying instructors by their usernames, and creating the registrations automatically.